-madded trivai file
