+++ code/embeddings/Spacy.py
@@ -3,7 +3,6 @@ import spacy
 import os
 import sys
 from random import randint
-from datasets import load_dataset
 cwd = os.getcwd()
 sys.path.append(cwd + "/code")
 from pre_processing import pre_process
@@ -15,31 +14,6 @@ class SpacyEmbed:
     def embed(self, text):
         text = pre_process(text, stopwords=True, lemmitize=True, stemming=True)
         return self.nlp(text)
-        
-
-class TriviaQA:
-    def __init__(self):
-        self.trainData = load_dataset('trivia_qa', 'rc')['train']
-
-    def getRandomIndex(self):
-        idx = randint(0, len(self.trainData)-1)
-        while len(self.getContext(idx)) == 0:
-            idx = randint(0, len(self.trainData))
-        return idx
-    
-    def getRandomDoc(self):
-        idx = self.getRandomIndex()
-        context = self.getContext(idx)
-        return context[randint(0, len(context)-1)]
-
-    def getQuestion(self, idx):
-        return self.trainData[idx]['question']
-
-    def getContext(self, idx):
-        return self.trainData[idx]['search_results']['search_context']
-
-    def __len__(self):
-        return len(self.trainData)
 
 
 def getSimilarityScore(questIndex, random=True):
@@ -54,7 +28,6 @@ def getSimilarityScore(questIndex, random=True):
 
     docEmbedding = spacyEmbed.embed(doc)
     return questEmbedding.similarity(docEmbedding)
-
 
 
 def getRetriveResult(scores):
